+++ service/controller/managementcluster/resource.go
@@ -298,6 +298,7 @@ func newResources(config resourcesConfig) ([]resource.Interface, error) {
 			K8sClient:       config.K8sClient,
 			Logger:          config.Logger,
 			PasswordManager: passwordManager,
+			Installation:    config.Installation,
 			Provider:        config.Provider,
 		}
 
@@ -349,17 +350,15 @@ func newResources(config resourcesConfig) ([]resource.Interface, error) {
 		alertmanagerConfigResource,
 		heartbeatWebhookConfigResource,
 		alertmanagerWiringResource,
+		remoteWriteAPIEndpointConfigSecretResource,
+		remoteWriteIngressAuthResource,
+		remoteWriteIngressResource,
 		scrapeConfigResource,
 		prometheusResource,
 		verticalPodAutoScalerResource,
 		monitoringIngressResource,
 		heartbeatResource,
 		pvcResizeResource,
-	}
-
-	// We enable those resources on CAPI clusters only as the agent is not running there and vintage clusters do not have MC namespaces.
-	if key.IsCAPIManagementCluster(config.Provider) {
-		resources = append(resources, remoteWriteAPIEndpointConfigSecretResource, remoteWriteIngressAuthResource, remoteWriteIngressResource)
 	}
 
 	{