+++ service/controller/resource/monitoring/prometheus/resource.go
@@ -271,7 +271,7 @@ func toPrometheus(ctx context.Context, v interface{}, config Config) (metav1.Obj
 		},
 	}
 
-	if !key.IsInCluster(config.Installation, cluster) {
+	if !key.IsManagementCluster(config.Installation, cluster) {
 		// Workload cluster
 		prometheus.Spec.APIServerConfig = &promv1.APIServerConfig{
 			Host: fmt.Sprintf("https://%s", key.APIUrl(cluster)),
@@ -340,13 +340,12 @@ func toPrometheus(ctx context.Context, v interface{}, config Config) (metav1.Obj
 			},
 		}
 
-		allMonitorSelector := []metav1.LabelSelectorRequirement{}
-		if key.IsCAPIManagementCluster(config.Provider) {
-			// We do not discover the service monitors discovered by the agent running on the management cluster
-			allMonitorSelector = append(allMonitorSelector, metav1.LabelSelectorRequirement{
+		// We do not discover the service monitors discovered by the agent running on the management cluster
+		allMonitorSelector := []metav1.LabelSelectorRequirement{
+			{
 				Key:      "application.giantswarm.io/team",
 				Operator: metav1.LabelSelectorOpDoesNotExist,
-			})
+			},
 		}
 
 		// An empty label selector matches all objects.